:alien: Code review and created project details component
created component and code review
