+++ src/app/components/ProjectDetailsCard/style.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+
+const Container = styled.div`
+  @media screen and (min-width: 1024px) {
+    width: 100vw;
+  }
+
+  @media screen and (min-width: 768px) {
+    width: 100vw;
+  }
+
+  @media screen and (min-width: 320px) {
+    width: 100vw;
+  }
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`
+
+export { Container }
