+++ src/app/components/index.js
@@ -5,3 +5,4 @@ export { default as ProjectCard } from './ProjectCard'
 export { default as ProjectList } from './ProjectList'
 export { default as SearchBox } from './SearchBox'
 export { default as StyledButton } from './Button'
+export { default as ProjectDetailsCard } from './ProjectDetailsCard'
