+++ src/app/pages/ProjectDetail.js
@@ -1,51 +1,11 @@
-import React, { useEffect, useState } from 'react'
-import { useParams } from 'react-router-dom'
-import { Footer, NavBar } from '../components'
-import axios from 'axios'
+import React from 'react'
+import { Footer, NavBar, ProjectDetailsCard } from '../components'
 
-function ProjectDetail () {
-  const { projectname, reponame } = useParams()
-  const [issues, setIssues] = useState([])
-
-  useEffect(() => {
-    projectFetch()
-  }, [])
-
-  const projectFetch = async () => {
-    const query = `{
-      issues(pathname:"/${projectname}/${reponame}") {
-        html_url
-        title
-        user{
-            avatar_url
-        }
-      }
-    }`
-
-    const res = await axios.get(`http://localhost:8080/graphql?query=${query}`)
-    // const res = await axios.get(`/graphql?query=${query}`)
-    setIssues(res.data.data.issues)
-  }
-
+const ProjectDetail = () => {
   return (
     <>
       <NavBar />
-      <div>
-        Total issue: {issues.length}
-        {issues.map((item, i) => (
-          <div key={i}>
-            <a href={item.html_url}>
-              <h2>{item.title}</h2>
-            </a>
-            <img
-              height="50px"
-              width="50px"
-              src={item.user.avatar_url}
-              alt="The issue creator"
-            />
-          </div>
-        ))}
-      </div>
+      <ProjectDetailsCard />
       <Footer />
     </>
   )
